third_party/WebKit/LayoutTests/fast/js/structured-clone.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
function promise_test(func, name, properties) {
    properties = properties || {};
    var test = async_test(name, properties);
    Promise.resolve(test.step(func, test, test))
           .then(function() { test.done(); })
           .catch(test.step_func(function(value) { throw value; }));
}

function structuredClone(o)
{
    return new Promise(function(resolve, reject) {
        var mc = new MessageChannel();
        mc.port2.onmessage = function(e) { resolve(e.data); };
        mc.port1.postMessage(o);
    });
}

promise_test(function() {
    var inner = {};
    var orig = { inner: inner };
    inner.outer = orig;
    return structuredClone(orig).then(function(clone) {
        assert_equals(clone.inner.outer, clone, 'Cycles should be preserved');
    });
}, 'Verify: "This algorithm preserves cycles..."');

promise_test(function() {
    var gen = {name: 'AES-CBC', length: 128};
    return crypto.subtle.generateKey(gen, false, ['encrypt']).then(function(key) {
        var simple = {};
        var blob = new Blob(['content']);
        var orig = {
            s1: simple, s2: simple,
            b1: blob, b2: blob,
            k1: key, k2: key
        };
        return structuredClone(orig).then(function(clone) {
            assert_equals(clone.s1, clone.s2, 'JS object identity should be preserved');
            assert_equals(clone.b1, clone.b2, 'Core object identity should be preserved');
            assert_equals(clone.k1, clone.k2, 'Module object identity should be preserved');
        });
    });
}, 'Verify: "This algorithm preserves... the identity of duplicate objects in graphs."');

promise_test(function() {
    var name = 'this is a setter';

    var orig = {};
    orig[name] = 'value';

    var setter_called = false;
    Object.defineProperty(Object.prototype, name, {
        set: function(a) {
            setter_called = true;
        }
    });

    assert_true(orig.hasOwnProperty(name));

    return structuredClone(orig).then(function(clone) {
        assert_equals(typeof clone, 'object', 'Clone should be an object');
        assert_false(setter_called, 'Setter should not be called by cloning algorithm.');
        assert_true(clone.hasOwnProperty(name), 'Cloning algorithm should add an own property.')
        assert_equals(clone[name], orig[name], 'Property value should match');
    });
}, 'Verify: "Add a new property..." (objects)');

promise_test(function() {
    var name = '123456';

    var orig = [];
    orig[name] = 'value';

    var setter_called = false;
    Object.defineProperty(Object.prototype, name, {
        set: function(a) {
            setter_called = true;
        }
    });

    assert_true(orig.hasOwnProperty(name));

    return structuredClone(orig).then(function(clone) {
        assert_true(Array.isArray(clone), 'Clone should be an Array');
        assert_false(setter_called, 'Setter should not be called by cloning algorithm.');
        assert_true(clone.hasOwnProperty(name), 'Cloning algorithm should add an own property.')
        assert_equals(clone[name], orig[name], 'Property value should match');
    });
}, 'Verify: "Add a new property..." (arrays)');

promise_test(function() {
    var name = '256';

    var orig = [];
    for (var i = 0; i < 256; ++i) {
        orig[i] = i;
    }

    orig[name] = 'value';

    var setter_called = false;
    Object.defineProperty(Object.prototype, name, {
        set: function(a) {
            setter_called = true;
        }
    });

    assert_true(orig.hasOwnProperty(name));

    return structuredClone(orig).then(function(clone) {
        assert_true(Array.isArray(clone), 'Clone should be an Array');
        assert_false(setter_called, 'Setter should not be called by cloning algorithm.');
        assert_true(clone.hasOwnProperty(name), 'Cloning algorithm should add an own property.')
        assert_equals(clone[name], orig[name], 'Property value should match');
    });
}, 'Verify: "Add a new property..." (dense arrays)');

</script>
